:recycle: [Refactor] App.tsx 1차 분리 #easy
